Refactor: Switch from a flat repo to a more structured repo.
