+++ include/sigma/enum.h
