+++ src/apex.cpp
@@ -1,0 +1,189 @@
+#include <algorithm>
+#include <cmath>
+
+#include "sigma/apex.h"
+
+ApexConverter::ApexConverter(std::vector<std::pair<std::uint8_t, double>> table)
+    : dectable_(std::move(table)), enctable_(dectable_) {
+  std::sort(enctable_.begin(), enctable_.end(),
+            [](auto &a, auto &b) { return a.second < b.second; });
+}
+
+std::optional<double> ApexConverter::decode_uint8(std::uint8_t code) const {
+  for (auto &kv : dectable_) {
+    if (kv.first == code)
+      return kv.second;
+  }
+  return std::nullopt;
+}
+
+std::uint8_t ApexConverter::encode_uint8(double val) const {
+  // nearest by absolute difference on value
+  const auto it = std::min_element(
+      enctable_.begin(), enctable_.end(), [val](const auto &a, const auto &b) {
+        return std::fabs(a.second - val) < std::fabs(b.second - val);
+      });
+  return it->first;
+}
+
+// --- Tables ---
+const ApexConverter ISOSpeedConverter({
+    {0, 6},       {3, 8},       {5, 10},       {8, 12},      {11, 16},
+    {13, 20},     {16, 25},     {19, 32},      {21, 40},     {24, 50},
+    {27, 64},     {29, 80},     {32, 100},     {35, 125},    {37, 160},
+    {40, 200},    {43, 250},    {45, 320},     {48, 400},    {51, 500},
+    {53, 640},    {56, 800},    {59, 1000},    {61, 1250},   {64, 1600},
+    {67, 2000},   {69, 2500},   {72, 3200},    {75, 4000},   {77, 5000},
+    {80, 6400},   {83, 8000},   {85, 10000},   {88, 12800},  {91, 16000},
+    {93, 20000},  {96, 25600},  {99, 32000},   {101, 40000}, {104, 51200},
+    {107, 64000}, {109, 80000}, {112, 102400},
+});
+
+const ApexConverter ExpComp2Converter({
+    {0, 0.0},
+    {4, 0.5},
+    {8, 1.0},
+    {12, 1.5},
+    {16, 2.0},
+    {20, 2.5},
+    {24, 3.0},
+    {232, -3.0},
+    {236, -2.5},
+    {240, -2.0},
+    {244, -1.5},
+    {248, -1.0},
+    {252, -0.5},
+});
+
+const ApexConverter ExpComp3Converter({
+    {0, 0.0},    {3, 0.3},    {5, 0.7},    {8, 1.0},    {11, 1.3},
+    {14, 1.7},   {16, 2.0},   {19, 2.3},   {21, 2.7},   {24, 3.0},
+    {27, 3.3},   {29, 3.7},   {32, 4.0},   {35, 4.3},   {37, 4.7},
+    {40, 5.0},   {43, 5.3},   {45, 5.7},   {48, 6.0},   {51, 6.3},
+    {205, -6.3}, {208, -6.0}, {211, -5.7}, {213, -5.3}, {216, -5.0},
+    {219, -4.7}, {221, -4.3}, {224, -4.0}, {227, -3.7}, {229, -3.3},
+    {232, -3.0}, {235, -2.7}, {237, -2.3}, {240, -2.0}, {243, -1.7},
+    {245, -1.3}, {248, -1.0}, {251, -0.7}, {253, -0.4},
+});
+
+const ApexConverter ShutterSpeed2Converter({
+    {17, 30},
+    {20, 20},
+    {24, 15},
+    {28, 10},
+    {32, 8},
+    {36, 6},
+    {40, 4},
+    {44, 3},
+    {48, 2},
+    {52, 1.5},
+    {56, 1},
+    {60, 0.7},
+    {64, 1.0 / 2},
+    {68, 1.0 / 3},
+    {72, 1.0 / 4},
+    {76, 1.0 / 6},
+    {80, 1.0 / 8},
+    {84, 1.0 / 10},
+    {88, 1.0 / 15},
+    {92, 1.0 / 20},
+    {96, 1.0 / 30},
+    {100, 1.0 / 45},
+    {104, 1.0 / 60},
+    {108, 1.0 / 90},
+    {112, 1.0 / 125},
+    {116, 1.0 / 180},
+    {120, 1.0 / 250},
+    {124, 1.0 / 350},
+    {128, 1.0 / 500},
+    {132, 1.0 / 750},
+    {136, 1.0 / 1000},
+    {140, 1.0 / 1500},
+    {144, 1.0 / 2000},
+    {148, 1.0 / 3000},
+    {152, 1.0 / 4000},
+    {156, 1.0 / 6000},
+    {160, 1.0 / 8000},
+    {168, 1.0 / 16000},
+    {176, 1.0 / 32000},
+});
+
+const ApexConverter ShutterSpeed3Converter({
+    {16, 30},
+    {19, 25},
+    {21, 20},
+    {24, 15},
+    {27, 13},
+    {29, 10},
+    {32, 8},
+    {35, 6},
+    {37, 5},
+    {40, 4},
+    {43, 3.2},
+    {45, 2.5},
+    {48, 2},
+    {51, 1.6},
+    {53, 1.3},
+    {56, 1},
+    {59, 0.8},
+    {61, 0.6},
+    {64, 0.5},
+    {67, 0.4},
+    {69, 0.3},
+    {72, 1.0 / 4},
+    {75, 1.0 / 5},
+    {77, 1.0 / 6},
+    {80, 1.0 / 8},
+    {83, 1.0 / 10},
+    {85, 1.0 / 13},
+    {88, 1.0 / 15},
+    {91, 1.0 / 20},
+    {93, 1.0 / 25},
+    {96, 1.0 / 30},
+    {99, 1.0 / 40},
+    {101, 1.0 / 50},
+    {104, 1.0 / 60},
+    {107, 1.0 / 80},
+    {109, 1.0 / 100},
+    {112, 1.0 / 125},
+    {115, 1.0 / 160},
+    {117, 1.0 / 200},
+    {120, 1.0 / 250},
+    {123, 1.0 / 320},
+    {125, 1.0 / 400},
+    {128, 1.0 / 500},
+    {131, 1.0 / 640},
+    {133, 1.0 / 800},
+    {136, 1.0 / 1000},
+    {139, 1.0 / 1250},
+    {141, 1.0 / 1600},
+    {144, 1.0 / 2000},
+    {147, 1.0 / 2500},
+    {149, 1.0 / 3200},
+    {152, 1.0 / 4000},
+    {155, 1.0 / 5000},
+    {157, 1.0 / 6000},
+    {160, 1.0 / 8000},
+    {163, 1.0 / 10000},
+    {165, 1.0 / 12800},
+    {168, 1.0 / 16000},
+    {171, 1.0 / 20000},
+    {173, 1.0 / 25600},
+    {176, 1.0 / 32000},
+});
+
+const ApexConverter Aperture2Converter({
+    {8, 1.0},  {12, 1.2}, {16, 1.4}, {20, 1.8}, {28, 2.5}, {32, 2.8}, {36, 3.5},
+    {40, 4.0}, {44, 4.5}, {48, 5.6}, {52, 6.7}, {56, 8.0}, {60, 9.5}, {64, 11},
+    {68, 13},  {72, 16},  {76, 19},  {80, 22},  {84, 27},  {88, 32},  {92, 38},
+    {96, 45},  {100, 54}, {104, 64}, {108, 76}, {112, 91},
+});
+
+const ApexConverter Aperture3Converter({
+    {8, 1.0},  {11, 1.1}, {13, 1.2}, {16, 1.4}, {19, 1.6}, {21, 1.8}, {24, 2.0},
+    {27, 2.2}, {29, 2.5}, {32, 2.8}, {35, 3.2}, {37, 3.5}, {40, 4.0}, {43, 4.5},
+    {45, 5.0}, {48, 5.6}, {51, 6.3}, {53, 7.1}, {56, 8.0}, {59, 9.0}, {61, 10},
+    {64, 11},  {67, 13},  {69, 14},  {72, 16},  {75, 18},  {77, 20},  {80, 22},
+    {83, 25},  {85, 29},  {88, 32},  {91, 36},  {93, 40},  {96, 45},  {99, 51},
+    {191, 57}, {104, 64}, {107, 72}, {108, 76}, {112, 91},
+});
